+++ app/src/main/java/com/surcumference/fingerprint/plugin/impl/qq/QQBasePlugin.java
@@ -1,4 +1,4 @@
-// Modified by mqmqgo, 2026-09-25: Keystore-only password check, removed telemetry/blacklist calls
+// Modified by mqmqgo, 2026-09-25: Keystore-only password check, removed telemetry/blacklist calls; password as wiped char[]
 package com.surcumference.fingerprint.plugin.impl.qq;
 
 import static com.surcumference.fingerprint.Constant.ICON_QQ_SETTING_ENTRY_DARK_BASE64;
@@ -41,6 +41,7 @@ import com.surcumference.fingerprint.util.DpUtils;
 import com.surcumference.fingerprint.util.ImageUtils;
 import com.surcumference.fingerprint.util.KeyboardUtils;
 import com.surcumference.fingerprint.util.StyleUtils;
+import com.surcumference.fingerprint.util.SecureChars;
 import com.surcumference.fingerprint.util.Task;
 import com.surcumference.fingerprint.util.ViewUtils;
 import com.surcumference.fingerprint.util.XBiometricIdentify;
@@ -335,7 +336,7 @@ public class QQBasePlugin implements IAppPlugin, IMockCurrentUser {
                 payDialog.withdrawTitleTextView.setText("使用指纹验证身份");
             }
             initFingerPrintLock(context, passwordEncrypted, (password) -> { // success
-                payDialog.inputEditText.setText(password);
+                payDialog.inputEditText.setText(password, 0, password.length);
                 if (longPassword) {
                     payDialog.okButton.performClick();
                 }
@@ -474,7 +475,8 @@ public class QQBasePlugin implements IAppPlugin, IMockCurrentUser {
                 .decryptPasscode(passwordEncrypted, new BizBiometricIdentify.IdentifyListener() {
 
                     @Override
-                    public void onDecryptionSuccess(BizBiometricIdentify identify, @NonNull String decryptedContent) {
+                    public void onDecryptionSuccess(BizBiometricIdentify identify, @NonNull char[] decryptedContent) {
+                        // decryptedContent is wiped by XBiometricIdentify right after this returns
                         super.onDecryptionSuccess(identify, decryptedContent);
                         onSuccessUnlockCallback.onFingerprintVerificationOK(decryptedContent);
                     }