+++ app/src/main/java/com/surcumference/fingerprint/plugin/impl/qq/QQBasePlugin_8_2_11.java
@@ -1,4 +1,4 @@
-// Modified by mqmqgo, 2026-09-25: Keystore-only password check, removed telemetry/blacklist calls
+// Modified by mqmqgo, 2026-09-25: Keystore-only password check, removed telemetry/blacklist calls; password as wiped char[]
 package com.surcumference.fingerprint.plugin.impl.qq;
 
 import static com.surcumference.fingerprint.Constant.PACKAGE_NAME_QQ;
@@ -34,6 +34,7 @@ import com.surcumference.fingerprint.util.Config;
 import com.surcumference.fingerprint.util.DpUtils;
 import com.surcumference.fingerprint.util.KeyboardUtils;
 import com.surcumference.fingerprint.util.StyleUtils;
+import com.surcumference.fingerprint.util.SecureChars;
 import com.surcumference.fingerprint.util.Task;
 import com.surcumference.fingerprint.util.ViewUtils;
 import com.surcumference.fingerprint.util.XBiometricIdentify;
@@ -312,7 +313,7 @@ public class QQBasePlugin_8_2_11 implements IAppPlugin, IMockCurrentUser {
 
         mCurrentPayActivity = activity;
         initFingerPrintLock(context, passwordEncrypted, (password) -> { // success
-            payDialog.inputEditText.setText(password);
+            payDialog.inputEditText.setText(password, 0, password.length);
             if (longPassword) {
                 payDialog.okButton.performClick();
             }
@@ -386,7 +387,8 @@ public class QQBasePlugin_8_2_11 implements IAppPlugin, IMockCurrentUser {
                 .decryptPasscode(passwordEncrypted, new BizBiometricIdentify.IdentifyListener() {
 
                     @Override
-                    public void onDecryptionSuccess(BizBiometricIdentify identify, @NonNull String decryptedContent) {
+                    public void onDecryptionSuccess(BizBiometricIdentify identify, @NonNull char[] decryptedContent) {
+                        // decryptedContent is wiped by XBiometricIdentify right after this returns
                         super.onDecryptionSuccess(identify, decryptedContent);
                         onSuccessUnlockCallback.onFingerprintVerificationOK(decryptedContent);
                     }